% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributorpRo.R
\name{contributorpRo}
\alias{contributorpRo}
\title{Retrieve All Contributors to Projects & Tasks}
\usage{
contributorpRo(client_id, secret)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}
}
\description{
This function allows you to easily retrieve users contributed to projects or tasks.
}
\examples{
taskpRo(client_id = client_id,
  secret = secret)
}
