% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDPpRo.R
\name{PDPpRo}
\alias{PDPpRo}
\title{Update DOMO Personalized Data Permission (PDP)}
\usage{
PDPpRo(client_id, secret, ds_id, parallel = FALSE)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}

\item{ds_id}{An alpha-numeric string that uniquely identifies a DOMO dataset, can be found on the address bar. The dataset must include `Dataset ID`, `Policy Name`, `Policy Column`, `User ID`, and `Policy Value` fields.}

\item{parallel}{Provides an option to use parallel feature from the "Foreach" package.}
}
\description{
This function allows you to easily add, update, and delete PDP policies based on a DOMO dataset
}
\examples{
PDPpRo(client_id = client_id,
  secret = secret,
  ds_id = '58dfaba9-c3c4-4099-b7ef-dc2ca222cb24',
  parallel = TRUE)
}
