% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskpRo.R
\name{taskpRo}
\alias{taskpRo}
\title{Retrieve All Tasks From All Projects}
\usage{
taskpRo(client_id, secret)
}
\arguments{
\item{client_id}{A client_id that can be created on the developer.domo.com page.}

\item{secret}{A secret that can created on the developer.domo.com page.}
}
\description{
This function allows you to easily retrieve all active lists & tasks from active projects.
}
\examples{
taskpRo(client_id = client_id,
  secret = secret)
}
